requirements.txt

fastapi>=0.103.0
uvicorn>=0.22.0
httpx>=0.24.1
pyjwt>=2.8.0
pydantic>=2.3.0
cryptography>=41.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
